Refactor - rename service name
due to amadeus developer service name change
