+++ src/main/java/org/routemaster/api/total/domain/air/service/AirInspirationService.java
@@ -1,4 +1,4 @@
 package org.routemaster.api.total.domain.air.service;
 
-public interface AirAIService {
+public interface AirInspirationService {
 }
