--- src/main/java/org/routemaster/api/total/domain/air/service/AirTravelInsightsService.java
@@ -1,4 +1,0 @@
-package org.routemaster.api.total.domain.air.service;
-
-public interface AirTravelInsightsService {
-}
