+++ src/main/java/org/routemaster/api/total/domain/air/service/impl/DefaultAirlinesService.java
@@ -5,9 +5,9 @@ import com.amadeus.Params;
 import com.amadeus.exceptions.ResponseException;
 import com.amadeus.resources.Destination;
 import lombok.extern.slf4j.Slf4j;
-import org.routemaster.api.total.infra.amadeus.vo.DestinationVO;
-import org.routemaster.api.total.domain.air.service.AirSearchAndShoppingService;
+import org.routemaster.api.total.domain.air.service.AirlinesService;
 import org.routemaster.api.total.infra.amadeus.value.AmadeusValue;
+import org.routemaster.api.total.infra.amadeus.vo.DestinationVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,41 +16,13 @@ import java.util.List;
 
 @Slf4j
 @Service
-public class DefaultAirSearchAndShoppingService implements AirSearchAndShoppingService {
+public class DefaultAirlinesService implements AirlinesService {
 
     private final Amadeus amadeus;
 
     @Autowired
-    public DefaultAirSearchAndShoppingService(AmadeusValue amadeusValue) {
+    public DefaultAirlinesService(AmadeusValue amadeusValue) {
         this.amadeus = amadeusValue.getAmadeus();
-    }
-
-    @Override
-    public List<DestinationVO> airportRoutes(String departureAirportCode, Long max) {
-        Params params = Params.with("departureAirportCode", departureAirportCode);
-
-        if (max != null) {
-            params.and("max", max);
-        }
-
-        return airportRoutes(params);
-    }
-
-    private List<DestinationVO> airportRoutes(Params params) {
-        try {
-            Destination[] destinations = amadeus.airport.directDestinations.get(params);
-            List<DestinationVO> destinationVOs = new ArrayList<>();
-            for (Destination destination : destinations) {
-                destinationVOs.add(
-                        DestinationVO.builder()
-                                .destination(destination)
-                                .build()
-                );
-            }
-            return destinationVOs;
-        } catch (ResponseException e) {
-            throw new RuntimeException(e); // Required Custom Exception Handling
-        }
     }
 
     @Override
